// src/generated/models/getSuggestionsStatus.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export type GetSuggestionsStatus = (typeof GetSuggestionsStatus)[keyof typeof GetSuggestionsStatus];

export const GetSuggestionsStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
} as const;